moved project build functionality to wash-lib
